fix  style

+++ www/app/entity/event.php
@@ -13,7 +13,7 @@ class Event extends \ZCL\DB\Entity
     public const TYPE_CUSTOMER = 1;
     public const TYPE_JOB = 2;
     public const TYPE_PAYMENT = 3;
-    
+
     protected function init() {
         $this->event_id = 0;
         $this->event_type = 0;
@@ -29,7 +29,7 @@ class Event extends \ZCL\DB\Entity
 
     public static function isNotClosedTask($user_id) {
         $conn = \ZCL\DB\DB::getConnect();
-        $cnt = Event::findCnt("isdone<>1 and event_type in(0,2) and user_id={$user_id} ");  
+        $cnt = Event::findCnt("isdone<>1 and event_type in(0,2) and user_id={$user_id} ");
         return $cnt;
 
     }
@@ -42,17 +42,17 @@ class Event extends \ZCL\DB\Entity
 
         return true;
     }
-    
+
     protected function afterLoad() {
         $this->eventdate = strtotime($this->eventdate);
         $this->createdon = strtotime($this->createdon);
 
         $xml = @simplexml_load_string($this->details);
 
- 
-        $this->amount = (double)($xml->amount[0]);
+
+        $this->amount = (float)($xml->amount[0]);
         $this->paytype = (int)($xml->paytype[0]);
 
         parent::afterLoad();
-    }     
+    }
 }
